refactor: Move files in the cdc root entry point
