--- integration-libs/cdc/core/auth/facade/index.ts
@@ -1,1 +1,0 @@
-export * from './cdc-auth.service';
