+++ integration-libs/cdc/core/auth/store/index.ts
@@ -1,2 +1,1 @@
-export * from './actions/index';
 export * from './effects/index';
