+++ integration-libs/cdc/root/facade/cdc-auth.service.spec.ts
@@ -12,7 +12,7 @@ import {
   UserIdService,
 } from '@spartacus/core';
 import { of } from 'rxjs';
-import { CdcAuthActions } from '../store';
+import { CdcAuthActions } from '../store/actions/index';
 import { CdcAuthService } from './cdc-auth.service';
 
 class MockAuthStorageService implements Partial<AuthStorageService> {
