+++ integration-libs/cdc/root/index.ts
@@ -1,3 +1,4 @@
 export * from './cdc-root.module';
 export * from './config/index';
 export * from './facade/index';
+export * from './store/actions/index';
